requirements.txt
omegaconf==2.2.3
hydra-core==1.2.0
tqdm==4.66.5
lightning==2.3.3
tensorboardX==2.6.2.2
numpy==1.24.3
tensorboard==2.14.0
pydantic==2.10.4
torch==2.1.1
torchvision==0.16.1
torchmetrics==1.5.2
packaging<25
six
